+++ src/main/java/ru/baldursgate3/tgbot/storeservice/dto/CharDto.java
@@ -1,0 +1,4 @@
+package ru.baldursgate3.tgbot.storeservice.dto;
+
+public record CharDto(Long id, String name) {
+}
