+++ src/main/java/ru/baldursgate3/tgbot/storeservice/entities/Character.java
@@ -1,0 +1,27 @@
+package ru.baldursgate3.tgbot.storeservice.entities;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Table(name = "characters")
+@Getter
+@Setter
+@EqualsAndHashCode(of = "id")
+public class Character {
+    private @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    Long id;
+    private String name;
+    @ManyToOne
+    @JoinColumn(name = "user_id",nullable = false)
+    private User user;
+}
