--- src/main/java/ru/baldursgate3/tgbot/storeservice/jpa/entities/User.java
@@ -1,56 +1,0 @@
-package ru.baldursgate3.tgbot.storeservice.jpa.entities;
-
-import jakarta.persistence.*;
-
-import java.util.List;
-import java.util.Objects;
-
-
-@Entity
-@Table(name = "users")
-public class User {
-    private @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long id;
-    private String name;
-    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL)
-    private List<Character> characterList;
-
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public List<Character> getCharacterList() {
-        return characterList;
-    }
-
-    public void setCharacterList(List<Character> characterList) {
-        this.characterList = characterList;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        User user = (User) o;
-        return Objects.equals(id, user.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-}
