+++ src/main/java/ru/baldursgate3/tgbot/storeservice/repositories/CharacterRepository.java
@@ -1,7 +1,7 @@
-package ru.baldursgate3.tgbot.storeservice.jpa.repositories;
+package ru.baldursgate3.tgbot.storeservice.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.baldursgate3.tgbot.storeservice.jpa.entities.Character;
+import ru.baldursgate3.tgbot.storeservice.entities.Character;
 
 public interface CharacterRepository extends JpaRepository<Character, Long> {
